more precise

+++ EXTRA/missingvalue.java
@@ -5,22 +5,22 @@ public class missingvalue {
         System.out.println("How many inputs are there?");
         int n = scanner.nextInt();
         scanner.nextLine(); //we should use nextLine after nextInt because int donot read \n character and stops before
-        float arrx[] = new float[n];
-        float arry[] = new float[n];
+        double arrx[] = new double[n];
+        double arry[] = new double[n];
         System.out.println("Enter the inputs(x)");
         for(int i=0;i<n;i++)
         {
-            arrx[i]=scanner.nextFloat();
+            arrx[i]=scanner.nextDouble();
         }
         System.out.println("Enter the inputs(y)");
         for(int i=0;i<n;i++)
         {
-            arry[i]=scanner.nextFloat();
+            arry[i]=scanner.nextDouble();
         }
-        float sumx=0;
-        float sumy=0;
-        float sumxy=0;
-        float sumx2=0;
+        double sumx=0;
+        double sumy=0;
+        double sumxy=0;
+        double sumx2=0;
         for(int i=0;i<n;i++)
         {
             sumx+=arrx[i];
@@ -28,7 +28,7 @@ public class missingvalue {
             sumxy+=(arrx[i]*arry[i]);
             sumx2+=(arrx[i]*arrx[i]);
         }
-        float a,b;
+        double a,b;
         b=((n*sumxy)-(sumx*sumy))/((n*sumx2)-(sumx*sumx));
         a=(sumy/n)-(b*(sumx/n));
         System.out.println("The coefficients are=> a:"+a+" and b:"+b);
@@ -38,15 +38,15 @@ public class missingvalue {
         if(choice.equals("Y") || choice.equals("y"))
         {
             System.out.println("Enter the value of x: ");
-            float x = scanner.nextFloat();
-            float y = a+(b*x);
+            double x = scanner.nextDouble();
+            double y = a+(b*x);
             System.out.println("Missing value y is equal to:"+y);
         }
         else
         {
             System.out.println("Enter the value of y: ");
-            float y = scanner.nextFloat();
-            float x = (y-a)/b;
+            double y = scanner.nextDouble();
+            double x = (y-a)/b;
             System.out.println("Missing value x is equal to:"+x);
         }
     }
